Reorganise again due to analyser error
Recommendation not to use Compile Analysers with Workspace Analysers - moved Generator analytics back to SourceGenerator Project
