+++ src/ReactiveUI.SourceGenerator.Tests/REACTIVE/ReactiveGeneratorTests.FromReactiveProperiesWithAttributes#TestVM.Properties.g.verified.cs
@@ -7,23 +7,23 @@ using ReactiveUI;
 
 namespace TestNs
 {
-	/// <summary>
-	/// Partial class for the TestVM which contains ReactiveUI Reactive property initialization.
-	/// </summary>
-	public partial class TestVM
-	{
-		[global::System.CodeDom.Compiler.GeneratedCode("ReactiveUI.SourceGenerators.ReactiveGenerator", "1.1.0.0")]
-		/// <inheritdoc cref="_test3"/>
-		[global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-		[System.Text.Json.Serialization.JsonInclude]
-		[System.Runtime.Serialization.DataMember]
-		public int Test3
-		{ 
-			get => _test3;
-			[global::System.Diagnostics.CodeAnalysis.MemberNotNull("_test3")]
-			set => this.RaiseAndSetIfChanged(ref _test3, value);
-		}
-	}
+    /// <summary>
+    /// Partial class for the TestVM which contains ReactiveUI Reactive property initialization.
+    /// </summary>
+    public partial class TestVM
+    {
+        [global::System.CodeDom.Compiler.GeneratedCode("ReactiveUI.SourceGenerators.ReactiveGenerator", "1.1.0.0")]
+        /// <inheritdoc cref="_test3"/>
+        [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+        [System.Text.Json.Serialization.JsonInclude]
+        [System.Runtime.Serialization.DataMember]
+        public int Test3
+        { 
+            get => _test3;
+            [global::System.Diagnostics.CodeAnalysis.MemberNotNull("_test3")]
+            set => this.RaiseAndSetIfChanged(ref _test3, value);
+        }
+    }
 }
 #nullable restore
 #pragma warning restore
