+++ src/ReactiveUI.SourceGenerator.Tests/REACTIVE/ReactiveGeneratorTests.FromReactiveProperties#TestVM.Properties.g.verified.cs
@@ -7,21 +7,21 @@ using ReactiveUI;
 
 namespace TestNs
 {
-	/// <summary>
-	/// Partial class for the TestVM which contains ReactiveUI Reactive property initialization.
-	/// </summary>
-	public partial class TestVM
-	{
-		[global::System.CodeDom.Compiler.GeneratedCode("ReactiveUI.SourceGenerators.ReactiveGenerator", "1.1.0.0")]
-		/// <inheritdoc cref="_test1"/>
-		[global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-		public int Test1
-		{ 
-			get => _test1;
-			[global::System.Diagnostics.CodeAnalysis.MemberNotNull("_test1")]
-			set => this.RaiseAndSetIfChanged(ref _test1, value);
-		}
-	}
+    /// <summary>
+    /// Partial class for the TestVM which contains ReactiveUI Reactive property initialization.
+    /// </summary>
+    public partial class TestVM
+    {
+        [global::System.CodeDom.Compiler.GeneratedCode("ReactiveUI.SourceGenerators.ReactiveGenerator", "1.1.0.0")]
+        /// <inheritdoc cref="_test1"/>
+        [global::System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+        public int Test1
+        { 
+            get => _test1;
+            [global::System.Diagnostics.CodeAnalysis.MemberNotNull("_test1")]
+            set => this.RaiseAndSetIfChanged(ref _test1, value);
+        }
+    }
 }
 #nullable restore
 #pragma warning restore
