+++ src/ReactiveUI.SourceGenerators/Diagnostics/SuppressionDescriptors.cs
