+++ src/ReactiveUI.SourceGenerators/Diagnostics/Suppressions/OAPHMethodDoesNotNeedToBeStaticDiagnosticSuppressor.cs
