+++ src/ReactiveUI.SourceGenerators/Diagnostics/Suppressions/ReactiveFieldDoesNotNeedToBeReadOnlyDiagnosticSuppressor.cs
@@ -41,7 +41,7 @@ namespace ReactiveUI.SourceGenerators.Diagnostics.Suppressions
 
                     // Check if the method is using [Reactive], in which case we should suppress the warning
                     if (declaredSymbol is IFieldSymbol fieldSymbol &&
-                        semanticModel.Compilation.GetTypeByMetadataName(AttributeDefinitions.ReactiveAttributeType) is INamedTypeSymbol reactiveSymbol &&
+                        semanticModel.Compilation.GetTypeByMetadataName("ReactiveUI.SourceGenerators.ReactiveAttribute") is INamedTypeSymbol reactiveSymbol &&
                         fieldSymbol.HasAttributeWithType(reactiveSymbol))
                     {
                         context.ReportSuppression(Suppression.Create(ReactiveFieldsShouldNotBeReadOnly, diagnostic));
