+++ src/ReactiveUI.SourceGenerators/ReactiveUI.SourceGenerators.csproj
@@ -29,8 +29,6 @@
       System.Index;
       System.Range;
     </PolySharpIncludeGeneratedTypes>
-
-    <IsPackable>false</IsPackable>
   </PropertyGroup>
   
   <!-- Include PolySharp to generate polyfills for all projects (on their .NET Standard 2.x targets) -->
@@ -49,6 +47,7 @@
   <!-- This ensures the library will be packaged as a source generator when we use `dotnet pack` -->
   <ItemGroup>
     <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
+    <None Include="$(OutputPath)\ReactiveUI.CodeFixes.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
   </ItemGroup>
 
   <!-- Include PolySharp generated code -->
